fix: bound durable session snapshot growth
