+++ pkg/services/factory_runtime/internal/services/orchestration/runtime/dispatch_worker_sessions_cutover.go
@@ -963,52 +963,3 @@ func historicalTranscriptResult(
 	}
 	return result, nil
 }
-func newRecordedWorkerSessionObservationWithRecording(
-	live workersessions.Service,
-	ledger recordings.RuntimeLedger,
-	projector factory.WorldStateProjector,
-	clock factory.Clock,
-	providerSessions providersessions.Service,
-	replayEvents []interfaces.FactoryEvent,
-	recordingID string,
-	recordingReader recordings.WorkerRecordingReader,
-	factorySessionIDs ...string,
-) workersessions.Service {
-	return newRecordedWorkerSessionObservationWithRestoredState(
-		live, ledger, projector, clock, providerSessions, replayEvents, recordingID,
-		recordingReader, nil, nil, factorySessionIDs...,
-	)
-}
-
-func newRecordedWorkerSessionObservationWithRestoredState(
-	live workersessions.Service,
-	ledger recordings.RuntimeLedger,
-	projector factory.WorldStateProjector,
-	clock factory.Clock,
-	providerSessions providersessions.Service,
-	replayEvents []interfaces.FactoryEvent,
-	recordingID string,
-	recordingReader recordings.WorkerRecordingReader,
-	restoredWorldState *interfaces.FactoryWorldState,
-	restoredEventPrefix []interfaces.FactoryEvent,
-	factorySessionIDs ...string,
-) workersessions.Service {
-	factorySessionID := ""
-	if len(factorySessionIDs) > 0 {
-		factorySessionID = strings.TrimSpace(factorySessionIDs[0])
-	}
-	return &recordedWorkerSessionObservation{
-		Service:             live,
-		ledger:              ledger,
-		durability:          completedFlushWatermarkReader(ledger),
-		projector:           projector,
-		clock:               clock,
-		providerSessions:    providerSessions,
-		replayEvents:        cloneAndSortFactoryEvents(replayEvents),
-		restoredWorldState:  restoredWorldState,
-		restoredEventPrefix: cloneFactoryEventsInOrder(restoredEventPrefix),
-		recordingID:         strings.TrimSpace(recordingID),
-		recordingReader:     recordingReader,
-		factorySessionID:    factorySessionID,
-	}
-}
