+++ pkg/services/factory_runtime/internal/services/orchestration/runtime/worker_session_recorded_usage.go
@@ -1,6 +1,7 @@
 package runtime
 
 import (
+	"context"
 	"fmt"
 	"sort"
 	"strconv"
@@ -9,6 +10,7 @@ import (
 
 	interfaces "github.com/portpowered/infinite-you/pkg/services/factory_definitions"
 	factory "github.com/portpowered/infinite-you/pkg/services/factory_runtime"
+	providersessions "github.com/portpowered/infinite-you/pkg/services/provider_sessions"
 	"github.com/portpowered/infinite-you/pkg/services/recordings"
 	"github.com/portpowered/infinite-you/pkg/services/work"
 	workersessions "github.com/portpowered/infinite-you/pkg/services/worker_sessions"
@@ -824,4 +826,161 @@ func recordedDispatchStateCursor(
 		}
 	}
 	return sequence, found
+}
+
+// cloneFactoryEventsInOrder keeps a detached copy without changing the
+// append-order prefix used to decide whether restored state is still current.
+func cloneFactoryEventsInOrder(events []interfaces.FactoryEvent) []interfaces.FactoryEvent {
+	if len(events) == 0 {
+		return nil
+	}
+	cloned := make([]interfaces.FactoryEvent, len(events))
+	for index, event := range events {
+		cloned[index] = event.Clone()
+	}
+	return cloned
+}
+
+func (s *recordedWorkerSessionObservation) projectRecordedWorldState(
+	ctx context.Context,
+	events []interfaces.FactoryEvent,
+	ordered []interfaces.FactoryEvent,
+	selectedTick int,
+) (interfaces.FactoryWorldState, error) {
+	if restored, ok := restoredWorldStateForEvents(s.restoredWorldState, s.restoredEventPrefix, events); ok {
+		return *restored, nil
+	}
+	if s == nil || s.projector == nil {
+		return interfaces.FactoryWorldState{}, workersessions.ErrObservationProjectionUnavailable
+	}
+	world, err := s.projector(ordered, selectedTick)
+	if err != nil {
+		return interfaces.FactoryWorldState{}, workersessions.ErrObservationProjectionUnavailable
+	}
+	return world, nil
+}
+
+func restoredWorldStateForEvents(
+	state *interfaces.FactoryWorldState,
+	prefix []interfaces.FactoryEvent,
+	events []interfaces.FactoryEvent,
+) (*interfaces.FactoryWorldState, bool) {
+	if state == nil || len(prefix) == 0 || len(events) < len(prefix) {
+		return nil, false
+	}
+	for index := range prefix {
+		if !sameFactoryEventIdentity(prefix[index], events[index]) {
+			return nil, false
+		}
+	}
+	for _, event := range events[len(prefix):] {
+		if factoryEventRequiresWorkerSessionProjection(*state, event) {
+			return nil, false
+		}
+	}
+	return state, true
+}
+
+func sameFactoryEventIdentity(left, right interfaces.FactoryEvent) bool {
+	if left.Id != "" || right.Id != "" {
+		return left.Id == right.Id
+	}
+	return left.Type == right.Type &&
+		left.Context.Tick == right.Context.Tick &&
+		left.Context.Sequence == right.Context.Sequence &&
+		left.Context.EventTime.Equal(right.Context.EventTime)
+}
+
+func factoryEventRequiresWorkerSessionProjection(
+	state interfaces.FactoryWorldState,
+	event interfaces.FactoryEvent,
+) bool {
+	switch event.Type {
+	case interfaces.FactoryEventTypeRunRequest,
+		interfaces.FactoryEventTypeInitialStructureRequest,
+		interfaces.FactoryEventTypeSessionStarted,
+		interfaces.FactoryEventTypeSessionLifecycleControl,
+		interfaces.FactoryEventTypeSessionPaused,
+		interfaces.FactoryEventTypeSessionResultUpdated,
+		interfaces.FactoryEventTypeSessionResumed,
+		interfaces.FactoryEventTypeSessionCompleted,
+		interfaces.FactoryEventTypeRunResponse:
+		return false
+	case interfaces.FactoryEventTypeWorkRequest:
+		return !restoredWorkRequestEventIsKnown(state, event)
+	default:
+		return true
+	}
+}
+
+func restoredWorkRequestEventIsKnown(
+	state interfaces.FactoryWorldState,
+	event interfaces.FactoryEvent,
+) bool {
+	workIDs := pointerStringSlice(event.Context.WorkIDs)
+	for _, workID := range workIDs {
+		if _, ok := state.WorkItemsByID[workID]; !ok {
+			return false
+		}
+	}
+	requestID := stringPointerValue(event.Context.RequestID)
+	if requestID != "" {
+		for key, request := range state.WorkRequestsByID {
+			if key == requestID || request.RequestID == requestID {
+				return len(workIDs) > 0 || len(request.WorkItems) > 0
+			}
+		}
+		return false
+	}
+	return len(workIDs) > 0
+}
+
+func newRecordedWorkerSessionObservationWithRecording(
+	live workersessions.Service,
+	ledger recordings.RuntimeLedger,
+	projector factory.WorldStateProjector,
+	clock factory.Clock,
+	providerSessions providersessions.Service,
+	replayEvents []interfaces.FactoryEvent,
+	recordingID string,
+	recordingReader recordings.WorkerRecordingReader,
+	factorySessionIDs ...string,
+) workersessions.Service {
+	return newRecordedWorkerSessionObservationWithRestoredState(
+		live, ledger, projector, clock, providerSessions, replayEvents, recordingID,
+		recordingReader, nil, nil, factorySessionIDs...,
+	)
+}
+
+func newRecordedWorkerSessionObservationWithRestoredState(
+	live workersessions.Service,
+	ledger recordings.RuntimeLedger,
+	projector factory.WorldStateProjector,
+	clock factory.Clock,
+	providerSessions providersessions.Service,
+	replayEvents []interfaces.FactoryEvent,
+	recordingID string,
+	recordingReader recordings.WorkerRecordingReader,
+	restoredWorldState *interfaces.FactoryWorldState,
+	restoredEventPrefix []interfaces.FactoryEvent,
+	factorySessionIDs ...string,
+) workersessions.Service {
+	factorySessionID := ""
+	if len(factorySessionIDs) > 0 {
+		factorySessionID = strings.TrimSpace(factorySessionIDs[0])
+	}
+	return &recordedWorkerSessionObservation{
+		Service:             live,
+		ledger:              ledger,
+		durability:          completedFlushWatermarkReader(ledger),
+		projector:           projector,
+		clock:               clock,
+		providerSessions:    providerSessions,
+		replayEvents:        cloneAndSortFactoryEvents(replayEvents),
+		restoredWorldState:  restoredWorldState,
+		restoredEventPrefix: cloneFactoryEventsInOrder(restoredEventPrefix),
+		recordingID:         strings.TrimSpace(recordingID),
+		recordingReader:     recordingReader,
+		factorySessionID:    factorySessionID,
+	}
 }
